[TIME-SLOT-WEEK-VIEW-COMP][VM|DS] Refactor to re-include typings for component to export.
