+++ src/time-slot-week-view/index.ts
@@ -1,1 +1,2 @@
 export * from "./time-slot-week-view";
+export * from "./types";
